+++ ChangeLog.md
@@ -29,8 +29,8 @@ All notable changes to this project will be documented in this file.
 - Maven download URL to "https://archive.apache.org/dist/maven/"
   instead of "https://dlcdn.apache.org/maven/maven-3/". Maven changed 
   the latter URL to provide the current version only.
-  This broke the Venice setup in demo mode and the installation of
-  additional 3rd party libs.
+  This unfortunately broke the Venice setup in demo mode and the 
+  installation of additional 3rd party libs.