+++ ChangeLog.md
@@ -15,7 +15,7 @@ All notable changes to this project will be documented in this file.
 - :openai module support for assistant run api (...)
 - :openai module support for assistant run steps api (...)
 
-- function `http-client-j8/create-authorization-bearer`
+- function `http-client-j8/create-authorization-bearer` and `http-client-j8/build-url`